requirement.txt
cffi==1.17.1
cryptography==45.0.4
et_xmlfile==2.0.0
lxml==5.4.0
msoffcrypto-tool==5.4.2
numpy==2.3.0
olefile==0.47
openpyxl==3.1.5
pandas==2.3.0
pip==25.1.1
pycparser==2.22
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
setuptools==75.5.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2